Added updated caliper class + new library jars
Made an updated caliper program, tried with a normal java application just because the maven version kept downloading files before it ran. Integrating it to a maven project or intellij shouldn't be a problem.

Just getting a java.lang.NoClassDefFoundError that spent time trying to fix but was unable to last night, will spend time today to sort that out.